moved adverts query to frontend-api package
